requirements.txt
cassandra-driver==3.7.0
docopt==0.6.2
pylru==1.0.9
